Enable WASI feature, enhance security and add SGX sample (#142)

+++ core/iwasm/runtime/vmcore-wasm/wasm_interp.h
@@ -55,7 +55,8 @@ typedef struct WASMInterpFrame {
 static inline unsigned
 wasm_interp_interp_frame_size(unsigned all_cell_num)
 {
-  return align_uint(offsetof(WASMInterpFrame, lp) + all_cell_num * 5, 4);
+  return align_uint((uint32)offsetof(WASMInterpFrame, lp)
+                    + all_cell_num * 5, 4);
 }
 
 void
